[IC-111] refactor: Severity 관련 로직 메서드 분리 (#93)
